[FIX] libc:  Actually make clock_t 64 bits
In January we defined clock() to return microseconds but left clock_t as an
unsigned long (32 bits).  This fixes calls to nanosleep() (and its wrappers
sleep(), doze(), etc.) waiting forever.  It's likely that select() was also
affected similarly and clock-based measurements were sometimes understated.

+++ lamp/POSIX-headers/machine/_types.h
@@ -22,7 +22,7 @@ typedef unsigned short      __uint16_t;
 typedef unsigned int        __uint32_t;
 typedef unsigned long long  __uint64_t;
 
-typedef unsigned long __clock_t;
+typedef unsigned long long __clock_t;
 
 typedef long __time_t;
